requirements.txt
pytest==7.4.3
allure-pytest==2.13.2
faker==21.0.0
json-server==0.1.3
requests==2.31.0
coverage==7.3.4